* Fix output upload * Add pre-remove hook support * Better code separation
