+++ executors/curved/dockerworker/worker/job.py
@@ -1,23 +1,21 @@
 import os
 import time
 import shutil
-import logging
 import traceback
 
 import harbor
 
+import util
+import logic
+
 from ..config import config
 from ..log import logger
-
-from backend import copy_from_backend, copy_to_backend
-
-DELAY_WAIT_DOCKER_MIN = 0.1
-DELAY_WAIT_DOCKER_MAX = 10.0
 
 
 def do_docker_job(job):
     logger.debug("Got job #{} with descriptor: {}".format(job.job_id, job.descriptor))
     try:
+        job.update_status("running")
         process(job)
         job.update_status("completed")
         logger.debug("Finished job #{}".format(job.job_id))
@@ -31,172 +29,21 @@ def do_docker_job(job):
         raise e
 
 
-def getargs(arg_dict, subst):
-    """
-    transofrm dict of arguments (kw and positional) to string of arguments,
-    and substitute $VARS from subst dict
-    """
-    args = []
-    kw_keys = [k for k in arg_dict.keys() if k.startswith('-')]
-    for k in kw_keys:
-        v = arg_dict[k]
-        if v is True:  # TODO: take care of False value or not?
-            args.append(k)
-        elif k.startswith('--'):
-            args.append("%s=%s" % (k, str(v)))
-        else:
-            args.append("%s %s" % (k, str(v)))
-    pos_keys = sorted([k for k in arg_dict.keys() if k.startswith("__POS")])
-    for k in pos_keys:
-        v = arg_dict[k]
-        if type(v) == str or type(v) == unicode:
-            args.append(v)
-        else:
-            args.extend(v)
-    s = ' '.join(args)
-    for k, v in subst.iteritems():
-        s = s.replace(k, str(v))
-    return s
-
-
-def descriptor_correct(job):
-    keys_needed = [
-        ('cmd', unicode),
-        ('args', dict) ,
-        ('cpu_per_container', int),
-        ('max_memoryMB', int),
-        ('min_memoryMB', int),
-        ('env_container', dict)
-    ]
-
-    for key, keytype in keys_needed:
-        assert isinstance(job.descriptor.get(key), keytype)
-
-
-def build_command(job):
-    args = getargs(job.descriptor['args'],
-        {
-            "$DATA_DIR": "/data",
-            "$OUTPUT_DIR": "/output",
-            "$INPUT_DIR": "/input",
-            "$JOB_ID": job.job_id,
-            "$TIMEHASH": hash(time.time())
-        }
-    )
-
-    command = "\"{cmd} {args}\"".format(cmd=job.descriptor['cmd'], args=args)
-
-    return command
-
-
-def create_containers(job, in_dir, out_dir):
-    # Add needed containers
-    needed = job.descriptor['env_container'].get('needed_containers') or []
-
-    logger.debug("Creating containers for job #{}".format(job.job_id))
-
-    mounted_names = []
-    for i, container in enumerate(needed):
-        image, volumes = container['name'], container['volumes']
-        assert isinstance(volumes, list)
-
-        if not config.ONLY_LOCAL_IMAGES:
-            harbor.pull_image(image)
-
-        tag = "JOB-{}-CNT-{}".format(job.job_id, i)
-        mounted_names.append(tag)
-
-        harbor.run(
-            image,
-            volumes=volumes,
-            detach=True,
-            name=tag,
-            command="echo {} app".format(tag)
-        )
-
-    # Execute environment container
-    if not config.ONLY_LOCAL_IMAGES:
-        harbor.pull_image(job.descriptor['env_container']['name'])
-
-    volumes = [
-        "{}:/input".format(in_dir),
-        "{}:/output".format(out_dir),
-    ]
-    command = build_command(job)
-
-    logger.debug('Executing: {}'.format(command))
-
-    return harbor.run(
-        job.descriptor['env_container']['name'],
-        working_dir=job.descriptor['env_container']['workdir'],
-        command=command,
-        volumes_from=mounted_names,
-        volumes=volumes,
-        detach=True,
-        binds={
-           in_dir:{'bind': '/input', 'ro': True},
-           out_dir:{'bind': '/output', 'ro': False},
-        }
-    )
-
-
-def create_workdir(job):
-    job_workdir = os.path.join(config.WORK_DIR, job.job_id)
-    os.mkdir(job_workdir)
-
-    input_dir  = os.path.join(job_workdir, "input")
-    os.mkdir(input_dir)
-
-    output_dir = os.path.join(job_workdir, "output")
-    os.mkdir(output_dir)
-
-    return job_workdir, input_dir, output_dir
-
-
-def get_input_files(job, in_dir):
-    for input_file in job.input:
-        logger.debug("job #{}: Download input {}".format(job.job_id, input_file))
-        copy_from_backend(input_file, in_dir)
-
-
-def upload_output_files(job, out_dir):
-    upload_uri = job.descriptor.get('output_uri') # should contain $JOB_ID
-    upload_uri = upload_uri.replace('$JOB_ID', job.job_id)
-
-    for output_file in os.listdir(out_dir):
-        output_file_path = os.path.join(out_dir, output_file)
-
-        if not os.path.isfile(output_file_path):
-            continue
-
-        logger.debug("job #{}: Upload file `{}` to `{}`".format(job.job_id, output_file, upload_uri))
-        copy_to_backend(output_file_path, upload_uri)
-
-
-def write_std_output(container_id, out_dir):
-    with open(os.path.join(out_dir, "stdout"), "w") as stdout_f:
-        for logline in harbor.logs(container_id, stdout=True, stderr=False, stream=True):
-            stdout_f.write(logline)
-
-    with open(os.path.join(out_dir, "stderr"), "w") as stderr_f:
-        for logline in harbor.logs(container_id, stdout=False, stderr=True, stream=True):
-            stderr_f.write(logline)
-
-
-
 def process(job):
-    descriptor_correct(job)
+    util.descriptor_correct(job)
 
-    job_dir, in_dir, out_dir = create_workdir(job)
+    job_dir, in_dir, out_dir = logic.create_workdir(job)
 
-    get_input_files(job, in_dir)
-    container_id = create_containers(job, in_dir, out_dir)
+    logic.get_input_files(job, in_dir)
+    container_id = logic.create_containers(job, in_dir, out_dir)
 
     while harbor.is_running(container_id):
         logger.debug("Container is running. Sleeping for {} sec.".format(config.SLEEP_TIME))
         time.sleep(config.SLEEP_TIME)
 
-    write_std_output(container_id, out_dir)
+    logic.write_std_output(container_id, out_dir)
 
-    upload_output_files(job, out_dir)
+    logic.upload_output_files(job, out_dir)
+
+    logic.pre_remove_hook()
     shutil.rmtree(job_dir)
