+++ executors/curved/example.cfg
@@ -16,3 +16,5 @@ ONLY_LOCAL_IMAGES = False
 
 DOCKER_URL = "unix://var/run/docker.sock"
 DOCKER_API_VERSION = "1.17"
+
+PRE_REMOVE_HOOK = ""
